[DCGEO-218] Move files to shared-utils module
