--- apps/bfd-common/bfd-common-util/src/main/java/gov/cms/bfd/common/exceptions/GeneratorException.java
@@ -1,12 +1,0 @@
-package gov.cms.bfd.common.exceptions;
-
-public class GeneratorException extends RuntimeException {
-
-  public GeneratorException(String message) {
-    super(message);
-  }
-
-  public GeneratorException(String message, Throwable cause) {
-    super(message, cause);
-  }
-}
