--- apps/bfd-common/bfd-common-util/src/main/java/gov/cms/bfd/common/exceptions/ParsingException.java
@@ -1,8 +1,0 @@
-package gov.cms.bfd.common.exceptions;
-
-public class ParsingException extends RuntimeException {
-
-  public ParsingException(String message) {
-    super(message);
-  }
-}
