--- apps/bfd-common/bfd-common-util/src/main/java/gov/cms/bfd/common/generators/token/parser/TokenParser.java
@@ -1,9 +1,0 @@
-package gov.cms.bfd.common.generators.token.parser;
-
-import gov.cms.bfd.common.generators.token.pattern.TokenPattern;
-import java.util.Queue;
-
-public interface TokenParser {
-
-  TokenPattern parse(Queue<Character> patternStream);
-}
