--- apps/bfd-common/bfd-common-util/src/main/java/gov/cms/bfd/common/generators/token/pattern/TokenPattern.java
@@ -1,74 +1,0 @@
-package gov.cms.bfd.common.generators.token.pattern;
-
-import gov.cms.bfd.common.exceptions.GeneratorException;
-import gov.cms.bfd.common.exceptions.ParsingException;
-import java.math.BigInteger;
-import java.util.Random;
-import lombok.Getter;
-
-public abstract class TokenPattern {
-
-  @Getter private BigInteger totalPermutations;
-
-  public void init() {
-    totalPermutations = calculatePermutations();
-  }
-
-  public String createToken() {
-    BigInteger randomNumber;
-
-    do {
-      randomNumber = new BigInteger(totalPermutations.bitLength(), new Random());
-    } while (randomNumber.compareTo(totalPermutations) >= 0);
-
-    return createToken(randomNumber);
-  }
-
-  public String createToken(long seed) {
-    return createToken(new BigInteger(String.valueOf(seed)));
-  }
-
-  public String createToken(BigInteger seed) {
-    try {
-      if (seed.compareTo(BigInteger.ZERO) < 0) {
-        throw new GeneratorException("Seed value can not be negative.");
-      } else if (totalPermutations.compareTo(seed) <= 0) {
-        throw new GeneratorException("Seed value is too high for the given generator");
-      }
-
-      return generateToken(seed);
-    } catch (GeneratorException e) {
-      // This will bubble all the way up so we can check what seed value caused failure and
-      // investigate.
-      throw new GeneratorException("Failed with seed: " + seed, e);
-    }
-  }
-
-  public String convertToken(String token, TokenPattern basePattern) {
-    BigInteger value = basePattern.parseTokenValue(token);
-
-    if (totalPermutations.compareTo(value) < 0) {
-      throw new ParsingException("Target pattern does not have enough permutations to convert.");
-    }
-
-    return createToken(value);
-  }
-
-  public BigInteger parseTokenValue(String tokenValue) {
-    if (tokenValue.length() != tokenLength()) {
-      throw new ParsingException("Given token value is invalid for this pattern");
-    }
-
-    return calculateTokenValue(tokenValue);
-  }
-
-  public abstract boolean isValidPattern(String value);
-
-  abstract String generateToken(BigInteger seed);
-
-  abstract BigInteger calculateTokenValue(String tokenString);
-
-  abstract int tokenLength();
-
-  abstract BigInteger calculatePermutations();
-}
