+++ apps/bfd-shared-utils/pom.xml
@@ -20,18 +20,19 @@
 				<!-- Just useful. -->
 				<groupId>com.google.guava</groupId>
 				<artifactId>guava</artifactId>
-				<version>24.1.1-jre</version>
+				<version>29.0-jre</version>
 			</dependency>
-
+		<!-- Compile time only -->
+		<dependency>
+			<groupId>org.projectlombok</groupId>
+			<artifactId>lombok</artifactId>
+		</dependency>
+		<!-- Testing -->
+		<dependency>
+			<groupId>junit</groupId>
+			<artifactId>junit</artifactId>
+			<scope>test</scope>
+		</dependency>
 	</dependencies>
-
-	<build>
-		<pluginManagement>
-			<plugins>
-			</plugins>
-		</pluginManagement>
-		<plugins>
-		</plugins>
-	</build>
 
 </project>
