+++ apps/bfd-shared-utils/src/main/java/gov/cms/bfd/sharedutils/generators/exceptions/GeneratorException.java
@@ -1,0 +1,12 @@
+package gov.cms.bfd.sharedutils.generators.exceptions;
+
+public class GeneratorException extends RuntimeException {
+
+  public GeneratorException(String message) {
+    super(message);
+  }
+
+  public GeneratorException(String message, Throwable cause) {
+    super(message, cause);
+  }
+}
