+++ apps/bfd-shared-utils/src/main/java/gov/cms/bfd/sharedutils/generators/exceptions/ParsingException.java
@@ -1,0 +1,8 @@
+package gov.cms.bfd.sharedutils.generators.exceptions;
+
+public class ParsingException extends RuntimeException {
+
+  public ParsingException(String message) {
+    super(message);
+  }
+}
