+++ apps/bfd-shared-utils/src/main/java/gov/cms/bfd/sharedutils/generators/token/parser/TokenParser.java
@@ -1,0 +1,9 @@
+package gov.cms.bfd.sharedutils.generators.token.parser;
+
+import gov.cms.bfd.sharedutils.generators.token.pattern.TokenPattern;
+import java.util.Queue;
+
+public interface TokenParser {
+
+  TokenPattern parse(Queue<Character> patternStream);
+}
