+++ apps/pom.xml
@@ -63,7 +63,6 @@
 		<module>bfd-pipeline</module>
 		<module>bfd-server</module>
 		<module>bfd-server-test-perf</module>
-		<module>bfd-common</module>
 		<!-- <module>bfd-server-test-functions</module> -->
 	</modules>
